+++ src/frontend/js/utils/errors/handle.ts
@@ -1,6 +1,28 @@
-// Generic error handler to be called whenever we need to do error reporting throughout the app
-// For now only logs the error but should upload the errors to something like Sentry later on
+import * as Sentry from '@sentry/browser';
+
+import { CommonDataProps } from 'types/commonDataProps';
+
+const context: CommonDataProps['context'] = (window as any)
+  .__richie_frontend_context__;
+
+if (context && context.sentry_dsn) {
+  Sentry.init({
+    dsn: context.sentry_dsn,
+    environment: context.environment,
+    release: RICHIE_VERSION,
+  });
+  Sentry.configureScope(scope => scope.setExtra('application', 'frontend'));
+}
+
+/**
+ * Generic error handler to be called whenever we need to do error reporting throughout the app.
+ * Passes errors to Sentry if available, logs the error to the console otherwise.
+ */
 // tslint:disable:no-console
 export const handle = (error: Error) => {
-  console.error(error);
+  if (context.sentry_dsn) {
+    Sentry.captureException(error);
+  } else {
+    console.error(error);
+  }
 };